Fix all Kotlin compilation errors from CI build
- CameraIntrinsicsProvider: add .toDouble() to Float branch of cx/cy so both
  branches return Double (was: Float vs Double type mismatch)
- CaptureMetricsStore/DepthGuidance: add minMeters, maxMeters, markerActive
  fields (with defaults) required by DepthDistanceOverlay and CaptureScreen
- CaptureController: add startPreview(), stopPreview(), setDistanceMarker(),
  clearDistanceMarker(), isMarkerSet() methods required by CaptureScreen
- CaptureController: use named parameter onSaved= in saveFrame() call to prevent
  trailing lambda binding to onError (the last param) instead of onSaved
- CaptureController: pass minMeters, maxMeters, markerActive into updateDepth()

+++ app/src/main/java/com/yourorg/objectcapture/camera/CameraIntrinsicsProvider.kt
@@ -38,12 +38,12 @@ class CameraIntrinsicsProvider @Inject constructor(
         val fy = (focalMm / sensorHeightMm) * size.height
 
         val cx = if (activeArray != null) {
-            (activeArray.exactCenterX() / activeArray.width()) * size.width
+            ((activeArray.exactCenterX() / activeArray.width()) * size.width).toDouble()
         } else {
             size.width / 2.0
         }
         val cy = if (activeArray != null) {
-            (activeArray.exactCenterY() / activeArray.height()) * size.height
+            ((activeArray.exactCenterY() / activeArray.height()) * size.height).toDouble()
         } else {
             size.height / 2.0
         }
